training/pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "kaggle>=1.7.4.5",
    "mlflow>=3.9.0",
    "numpy<2",
    "opencv-python-headless<4.12",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.253",
    "wandb>=0.24.0",
    "boto3>=1.34.0",
    "apache-airflow>=2.8.0",
]
